+++ Sources/POCMoney/Utils/NonZeroInt/NonZeroInt.swift
@@ -1,7 +1,7 @@
 struct NonZeroInt: Equatable, Hashable, Sendable {
-    let rawValue: Int
+    let rawValue: Int64
 
-    init?(_ value: Int) {
+    init?(_ value: Int64) {
         guard value != 0 else {
             return nil
         }
@@ -10,7 +10,7 @@ struct NonZeroInt: Equatable, Hashable, Sendable {
     }
 
     /// Returns `-1` if this value is negative and `1` if it's positive.
-    var signum: Int {
+    var signum: Int64 {
         rawValue < 0 ? -1 : 1
     }
 }
@@ -19,7 +19,7 @@ extension NonZeroInt {
     /// Returns the quotient and remainder of this value divided by the given
     /// number of parts.
     ///
-    /// The remainder comes back as a `Remainder` rather than an `Int`, so a
+    /// The remainder comes back as a `Remainder` rather than an `Int64`, so a
     /// caller has to decide what to do when it is non-zero instead of being
     /// able to ignore it.
     ///
@@ -33,8 +33,8 @@ extension NonZeroInt {
     ///   `Remainder`. A non-zero remainder has the same sign as this value.
     func quotientAndRemainder(
         dividingBy rhs: PartCount
-    ) -> (quotient: Int, remainder: Remainder) {
-        let (quotient, remainder) = rawValue.quotientAndRemainder(dividingBy: Int(rhs))
+    ) -> (quotient: Int64, remainder: Remainder) {
+        let (quotient, remainder) = rawValue.quotientAndRemainder(dividingBy: Int64(rhs))
 
         return (quotient, Remainder(remainder))
     }
